requirements.txt
# Cell War Driver (cwd) dependencies

# Environment configuration
python-dotenv>=0.19.0

# Serial communication
pyserial>=3.5

# YAML parsing for smart configuration
pyyaml>=6.0

# Date/time handling
pytz>=2022.1

# Command-line interface
argparse>=1.4.0

# Type checking
typing-extensions>=4.0.0

# File and data processing
pandas>=1.4.0

# Optional dependencies
# For database support (if enabled with --use-database)
sqlalchemy>=1.4.0

# For Windows GUI launcher
# tkinter is typically bundled with Python installation
